# pkg/mixer/api/docs/swagger.yaml
basePath: /api/v1
definitions:
  api.ActionResponse:
    properties:
      name:
        description: Name of the rack (for import/approve).
        example: rack-01
        type: string
      status:
        description: Status of the action.
        example: activated
        type: string
    type: object
  api.ApproveRequest:
    properties:
      name:
        description: Name to assign to the rack (optional, defaults to current).
        example: rack-prod-01
        type: string
    type: object
...
  api.RackResponse:
    properties:
      config:
        allOf:
        - $ref: '#/definitions/config.RackConfig'
        description: Config (Runtime).
      first_seen:
        description: FirstSeen timestamp.
        example: "2023-10-26T10:00:00Z"
        type: string
      id:
        description: ID is the entity ID of the rack.
        example: "45612378"
        type: string
      ip:
        description: IP address of the rack agent.
        example: 192.168.1.50
        type: string
      last_seen:
        description: LastSeen timestamp.
        example: "2023-10-27T10:00:00Z"
        type: string
      machine_id:
        description: MachineID is the physical machine ID.
        example: 019de49b-0000-7000-8000-000000000000
        type: string
      name:
        description: Name of the rack.
        example: rack-01
        type: string
      port:
        description: Port number of the rack agent.
        example: 4222
        type: integer
      stats:
        allOf:
        - $ref: '#/definitions/api.RackStats'
        description: Stats (Transient/Snapshot).
      status:
        description: Status of the rack (active, inactive, pending).
        example: active
        type: string
    type: object
  api.RackStats:
    properties:
      cpu_usage:
        description: CPU usage percentage.
        example: 12.5
        type: number
      load_avg:
        description: Load average (1m).
        example: 0.5
        type: number
      mem_usage:
        description: Memory usage percentage.
        example: 45.2
        type: number
      uptime:
        description: Uptime in seconds.
        example: 3600
        type: integer
    type: object
  api.ScenarioImportResponse:
    properties:
      name:
        description: Name of the imported scenario.
        example: payment-switch-v1
        type: string
      status:
        description: Status of the import.
        example: imported_and_activated
        type: string
    type: object
...
  config.BaseConfig:
    properties:
      name:
        example: node-01
        type: string
      stateDir:
        example: ./data
        type: string
      stateFile:
        example: rack.flux
        type: string
    type: object
...
  config.LoggingConfig:
    properties:
      compress:
        example: true
        type: boolean
      debug:
        description: Hard-override for debug level
        type: boolean
      filename:
        example: logs/fluxrig.log
        type: string
      level:
        example: info
        type: string
      maxBackups:
        example: 7
        type: integer
      maxSizeMB:
        example: 100
        type: integer
      throttling:
        $ref: '#/definitions/config.ThrottlingConfig'
      trace:
        description: Hard-override for trace level
        type: boolean
    type: object
  config.MetricsConfig:
    properties:
      bentoEnabled:
        type: boolean
      hostEnabled:
        type: boolean
      runtimeEnabled:
        type: boolean
    type: object
...
  config.RackConfig:
    properties:
      base:
        $ref: '#/definitions/config.BaseConfig'
      logging:
        $ref: '#/definitions/config.LoggingConfig'
      rack:
        $ref: '#/definitions/config.RackSettings'
      snake:
        $ref: '#/definitions/config.SnakeConfig'
      store:
        $ref: '#/definitions/config.StoreConfig'
      telemetry:
        $ref: '#/definitions/config.TelemetryConfig'
    type: object
  config.RackSettings:
    properties:
      cleanupTimeout:
        type: string
      convergenceTimeout:
        type: string
      enrollmentInterval:
        type: string
      enrollmentTimeout:
        type: string
      handshakeInterval:
        type: string
      heartbeatInterval:
        type: string
      ip:
        example: 10.0.0.5
        type: string
      machineID:
        description: 'Runtime only: set via Passport or Mixer assignment'
        type: string
      maxHops:
        type: integer
      maxPayloadSize:
        type: integer
      namePrefix:
        type: string
    type: object
  config.SnakeConfig:
    properties:
      connectTimeout:
        example: 10s
        type: string
      convergenceDelay:
        example: 100ms
        type: string
      domain:
        example: flux
        type: string
      inactiveThreshold:
        example: 30s
        type: string
      insecureSkipVerify:
        type: boolean
      operationTimeout:
        example: 5s
        type: string
      port:
        example: 4222
        type: integer
      reconnectWait:
        example: 1s
        type: string
      rootCAFile:
        type: string
      streamName:
        example: flux-msg
        type: string
      streamSubjects:
        items:
          type: string
        type: array
      subscriptionRetryAttempts:
        example: 5
        type: integer
      subscriptionRetryWait:
        example: 200ms
        type: string
      tlscertFile:
        type: string
      tlskeyFile:
        type: string
      url:
        example: nats://localhost:4222
        type: string
    type: object
  config.StoreConfig:
    properties:
      clusterKeyFile:
        example: cluster.key
        type: string
      databaseFile:
        example: flux.duckdb
        type: string
      dir:
        example: ./data
        type: string
      walmaxSizeMB:
        example: 500
        type: integer
    type: object
  config.TelemetryConfig:
    properties:
      baseSubject:
        type: string
      batchInterval:
        type: string
      disabled:
        description: Master switch to disable all telemetry
        type: boolean
      maxBatchSize:
        type: integer
      metrics:
        $ref: '#/definitions/config.MetricsConfig'
      serviceName:
        type: string
      streamName:
        type: string
      streamSubjects:
        items:
          type: string
        type: array
    type: object
  config.ThrottlingConfig:
    properties:
      burst:
        example: 50
        type: integer
      enabled:
        example: true
        type: boolean
      rate:
        example: 500
        type: number
    type: object
  registry.GearSpec:
    properties:
      config:
        additionalProperties:
          type: string
        type: object
      deploy:
        description: string (Group/Rack) or map
        example: rack-group-1
        type: string
      doc:
        example: Primary Ingress
        type: string
      id:
        description: Assigned by Registry/Mixer
        example: "100"
        type: string
      name:
        example: iso8583-in
        type: string
      ports:
        additionalProperties:
          type: string
        description: Assigned by Mixer
        type: object
      type:
        example: iso8583-server
        type: string
    type: object
  registry.LabelMatch:
    properties:
      labels:
        additionalProperties:
          type: string
        type: object
    type: object
  registry.RackTarget:
    properties:
      config:
        additionalProperties: {}
        type: object
      defaults:
        additionalProperties: {}
        type: object
      group:
        type: string
      labels:
        additionalProperties:
          type: string
        type: object
      match:
        $ref: '#/definitions/registry.LabelMatch'
      name:
        type: string
    type: object
  registry.Scenario:
    properties:
      gears:
        items:
          $ref: '#/definitions/registry.GearSpec'
        type: array
      meta:
        $ref: '#/definitions/registry.ScenarioMeta'
      racks:
        items:
          $ref: '#/definitions/registry.RackTarget'
        type: array
      wires:
        items:
          $ref: '#/definitions/registry.WireSpec'
        type: array
    type: object
  registry.ScenarioMeta:
    properties:
      name:
        example: payment-switch
        type: string
      version:
        example: 1.0.0
        type: string
    type: object
  registry.WireSpec:
    properties:
      from:
        example: iso8583-in.out
        type: string
      id:
        description: Assigned by Registry/Mixer
        type: string
      to:
        example: router.in
        type: string
    type: object
...
host: localhost:8090
info:
  contact:
    email: fluxrig@jaab.tech
    name: JAAB Tech Support
    url: https://fluxrig.org
  description: Control Plane API for fluxrig Orchestration. The Mixer acts as the
    brain of the fleet, managing Rack enrollment, Scenario deployment, and global
    Telemetry aggregation.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: fluxrig Mixer API
  version: v0.5.0-dev
paths:
...
  /entities/stats:
    get:
      description: Returns cached metrics for a specific entity or all entities.
      parameters:
      - description: Entity ID
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Get Entity Stats
      tags:
      - telemetry
...
  /racks:
    get:
      description: Returns a list of all known Racks
      parameters:
      - description: Filter by status (active, pending, offline)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api.RackResponse'
            type: array
      summary: List Racks
      tags:
      - registry
  /racks/{id}/{action}:
    post:
      consumes:
      - application/json
      description: 'Perform actions on Racks (Approve, Suspend, Activate, Remove,
        Log-Level). Note: This single endpoint handles multiple actions for simplicity
        in this Alpha version.'
      parameters:
      - description: Rack ID
        in: path
        name: id
        required: true
        type: string
      - description: Action (approve, suspend, activate)
        in: path
        name: action
        type: string
      - description: Approve Request Body
        in: body
        name: body
        schema:
          $ref: '#/definitions/api.ApproveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ActionResponse'
      summary: Rack Actions
      tags:
      - registry
...
  /scenario/import:
    post:
      consumes:
      - application/x-yaml
      description: Upload a YAML scenario definition to update the topology.
      parameters:
      - description: Validate without applying
        in: query
        name: dry_run
        type: boolean
      - description: Immediately activate after import
        in: query
        name: activate
        type: boolean
      - description: Scenario YAML
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/registry.Scenario'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ScenarioImportResponse'
      summary: Import Scenario
      tags:
      - scenario
  /telemetry/{type}:
    get:
      description: Query Logs or Metrics from DuckDB
      parameters:
      - description: Type of telemetry (logs, metrics)
        in: path
        name: type
        required: true
        type: string
      - description: Filter by Entity Name
        in: query
        name: entity
        type: string
      - description: Filter by Level (logs only)
        in: query
        name: min_level
        type: string
      - description: Start time (e.g. 1h, 2023-01-01T00:00:00Z)
        in: query
        name: since
        type: string
      - description: Max records (default 100)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      summary: Query Telemetry
      tags:
      - telemetry
...
schemes:
- http
swagger: "2.0"